# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Library Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  user: "user",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Book {
  id: number;
  title: string;
  author: string;
  category: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  isbn?: string | null;
  totalCopies: number;
  availableCopies: number;
  createdAt: string;
}

export interface BookInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  author: string;
  /** @minLength 1 */
  category: string;
  description?: string;
  isbn?: string;
  /** @minimum 1 */
  totalCopies: number;
}

export interface BookUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  author?: string;
  /** @minLength 1 */
  category?: string;
  description?: string;
  isbn?: string;
  /** @minimum 1 */
  totalCopies?: number;
}

export interface BorrowInput {
  bookId: number;
}

export type BorrowStatus = (typeof BorrowStatus)[keyof typeof BorrowStatus];

export const BorrowStatus = {
  issued: "issued",
  returned: "returned",
} as const;

export interface Borrow {
  id: number;
  userId: number;
  bookId: number;
  borrowDate: string;
  /** @nullable */
  returnDate?: string | null;
  /** @nullable */
  dueDate?: string | null;
  status: BorrowStatus;
  book?: Book;
  user?: User;
}

export interface DashboardStats {
  totalBooks: number;
  totalUsers: number;
  activeBorrows: number;
  totalBorrows: number;
  availableBooks: number;
  overdueBooks: number;
}

export type GetBooksParams = {
  search?: string;
  category?: string;
};