# lib/api-zod/src/generated/types/book.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Library Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Book {
  id: number;
  title: string;
  author: string;
  category: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  isbn?: string | null;
  totalCopies: number;
  availableCopies: number;
  createdAt: Date;
}