# lib/api-zod/src/generated/types/bookInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Library Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface BookInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  author: string;
  /** @minLength 1 */
  category: string;
  description?: string;
  isbn?: string;
  /** @minimum 1 */
  totalCopies: number;
}