# lib/api-zod/src/generated/types/borrow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Library Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { Book } from "./book";
import type { BorrowStatus } from "./borrowStatus";
import type { User } from "./user";

export interface Borrow {
  id: number;
  userId: number;
  bookId: number;
  borrowDate: Date;
  /** @nullable */
  returnDate?: Date | null;
  /** @nullable */
  dueDate?: Date | null;
  status: BorrowStatus;
  book?: Book;
  user?: User;
}