# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Library Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalBooks: number;
  totalUsers: number;
  activeBorrows: number;
  totalBorrows: number;
  availableBooks: number;
  overdueBooks: number;
}